+++ R/shortestroute_cleanup.R
@@ -62,7 +62,7 @@ removeduplicates <- function(rivers) {
 #' @importFrom graphics plot
 #' @importFrom graphics text
 #' @export
-cleanup <- function(rivers) {
+cleanup1 <- function(rivers) {
   if(!interactive()) stop("The cleanup() function can only be used in an interactive environment.")
   cat("Cleanup started, with",length(rivers$lines),"segments.",'\n')
   plot(rivers)
@@ -177,26 +177,35 @@ cleanup <- function(rivers) {
     rivers4<-rivers3
   }
   if(needed) {
-    cat("Segments must be split for connectedness to be correct.",'\n')
-    cat("Splitting segments...",'\n')
-    suppressMessages(rivers4 <- splitsegments(rivers=rivers3))
-    cat("Identified",(length(rivers4$lines)-length(rivers3$lines)),"new segment breaks.",'\n')
+    cat("Recommend splitting segments for connectedness to be correct.",'\n')  ############################
+    yes<-0
+    while(!any(yes==c("y","Y","n","N"))) yes <- readline(prompt="Split segments? (y/n) ")
+    if(yes %in% c("y","Y")) {
+      cat("Splitting segments...",'\n')
+      suppressMessages(rivers4 <- splitsegments(rivers=rivers3))
+      cat("Identified",(length(rivers4$lines)-length(rivers3$lines)),"new segment breaks.",'\n')
+    }
+    else rivers4 <- rivers3 ###########################
   }
   
   # checking segment lengths
+  cat("Checking between-vertex lengths...",'\n') ###########################
   distances <- NULL
+  if(is.null(rivers4$cumuldist)) rivers4 <- addcumuldist(rivers=rivers4)
   for(segi in 1:length(rivers4$lines)) {
-    if(dim(rivers4$lines[[segi]])[1]>1) {
-      for(verti in 2:(dim(rivers4$lines[[segi]])[1])) {
-        distances <- c(distances,pdist(rivers4$lines[[segi]][verti,],rivers4$lines[[segi]][(verti-1),]))
-      }
-    }
+    # if(dim(rivers4$lines[[segi]])[1]>1) {
+    #   for(verti in 2:(dim(rivers4$lines[[segi]])[1])) {
+    #     distances <- c(distances,pdist(rivers4$lines[[segi]][verti,],rivers4$lines[[segi]][(verti-1),]))
+    #   }
+    # }
+    ll <- length(rivers4$cumuldist[[segi]])
+    distances <- c(distances,(rivers4$cumuldist[[segi]][2:ll] - rivers4$cumuldist[[segi]][1:(ll-1)]))
   }
   maxlength <- max(distances,na.rm=T)
   cat("Maximum distance between vertices is",round(maxlength),'\n')
   hist(distances)
   yes <- 0
-  while(!any(yes==c("y","Y","n","N"))) yes <- readline(prompt="Insert vertices to reduce distances and increase point snapping precision? (y/n) ")
+  while(!any(yes==c("y","Y","n","N"))) yes <- readline(prompt="Insert vertices to reduce distances between vertices and increase point snapping precision? (y/n) ")
   if(yes=="Y" | yes=="y") {
     mindist <- as.numeric(readline(prompt="Minimum distance to use: "))
     cat("Inserting vertices...",'\n')
@@ -230,70 +239,258 @@ cleanup <- function(rivers) {
     cat('\n',"Checking for unconnected segments...",'\n')
     takeout <- NULL
     k <- 1
+    # 
+    # ################################## don't think i need to do this -- just makes sequential run faster
+    # dists <- rep(NA,length(rivers4$lines))
+    # for(i in 1:length(rivers4$lines)) {
+    #   dists[i] <- pdist(rivers4$lines[[rivers4$mouth$mouth.seg]][rivers4$mouth$mouth.vert,],
+    #                     c(mean(rivers4$lines[[i]][,1]),mean(rivers4$lines[[i]][,2])))
+    # }
+    # order <- order(dists)
+    # rivers41 <- rivers4
+    # rivers41$connections <- rivers4$connections[order,order]
+    # rivers41$lines <- rivers4$lines[order]
+    # origin <- rivers4$mouth$mouth.seg
+    # origin41 <-which(order==origin)
+    # 
+    # for(i in 1:length(rivers41$lines)) {   ### this is SOOO SLOOOOWWWW ... only check endpts??
+    #   if(is.na(detectroute1(end=origin41,start=i,rivers=rivers41,stopiferror=FALSE,algorithm="Dijkstra")[1])) {    #was sequential           ##################################################################
+    #     takeout[k] <- i
+    #     k <- k+1
+    #     # print(i)    ##############################################################################################
+    #   }
+    # }    # 
+    ################################## don't think i need to do this -- just makes sequential run faster
     
-    dists <- rep(NA,length(rivers4$lines))
-    for(i in 1:length(rivers4$lines)) {
-      dists[i] <- pdist(rivers4$lines[[rivers4$mouth$mouth.seg]][rivers4$mouth$mouth.vert,],
-                        c(mean(rivers4$lines[[i]][,1]),mean(rivers4$lines[[i]][,2])))
-    }
-    order <- order(dists)
-    rivers41 <- rivers4
-    rivers41$connections <- rivers4$connections[order,order]
-    rivers41$lines <- rivers4$lines[order]
-    origin <- rivers4$mouth$mouth.seg
-    origin41 <-which(order==origin)
+    # system.time({
+    # for(i in 1:length(rivers4$lines)) {   ### this is SOOO SLOOOOWWWW ... only check endpts??
+    #   if(is.na(detectroute1(end=rivers4$mouth$mouth.seg,start=i,rivers=rivers4,stopiferror=FALSE,algorithm="Dijkstra")[1])) {    #was sequential           ##################################################################
+    #     takeout[k] <- i
+    #     k <- k+1
+    #     # print(i)    ##############################################################################################
+    #   }
+    # }
+    # })   # 139.73 seconds
     
-    for(i in 1:length(rivers41$lines)) {
-      if(is.na(detectroute(end=origin41,start=i,rivers=rivers41,stopiferror=FALSE,algorithm="Dijkstra")[1])) {    #was sequential
-        takeout[k] <- i
-        k <- k+1
+    # system.time({
+      ntop <- rowSums(rivers4$connections==1,na.rm=T) + rowSums(rivers4$connections==2,na.rm=T) + rowSums(rivers4$connections==5,na.rm=T) + rowSums(rivers4$connections==6,na.rm=T)
+      nbot <- rowSums(rivers4$connections==3,na.rm=T) + rowSums(rivers4$connections==4,na.rm=T) + rowSums(rivers4$connections==5,na.rm=T) + rowSums(rivers4$connections==6,na.rm=T)
+      
+      thesefirst <- xor(ntop==0,nbot==0)
+      
+      checked <- takeout <- rep(F,length(rivers4$lines))
+      pb <- txtProgressBar(style=3)
+      while(!all(checked)) {   
+        # i <- which.min(checked) 
+        i <- ifelse(any(thesefirst),which.max(thesefirst),which.min(checked))
+        theroute <- detectroute1(end=rivers4$mouth$mouth.seg,start=i,rivers=rivers4,stopiferror=FALSE,algorithm="Dijkstra")
+        if(is.na(theroute[1])) { 
+          takeout[i] <- T
+          checked[i] <- T
+          thesefirst[i] <- F
+          # print(i)    ##############################################################################################
+        }
+        else {
+          checked[theroute] <- T
+          thesefirst[theroute] <- F
+        }
+        setTxtProgressBar(pb=pb, value=mean(checked))
       }
-    }
-    if(!is.null(takeout)) {
-      takeout <- order[takeout]
+      takeout <- which(takeout)
+    # })  # 76.19 # 84.46 seconds
+    
+    if(length(takeout)>0) {
+      # takeout <- order[takeout]   ##########################
       highlightseg(seg=takeout,rivers=rivers4)
-      cat(length(takeout),"unconnected segments detected.")
-      whattodo<-0
-      while(!any(whattodo==c("r","R","c","C"))) {
-        whattodo <- readline(prompt="Please select - (r)emove unconnected segments or (c)onnect segments & check again (r/c) ")
+      cat('\n',length(takeout),"unconnected segments detected.",'\n')
+      
+      ################## look for connected regions, make a list, deal with each region sequentially
+      # don't remove until at end
+      cat("Identifying regions...",'\n')
+      
+      num<-0
+      # system.time({
+      regions <- list()
+      regioni <- 1
+      toplace <- takeout
+      everybodyoutofthepool <- F
+      pb <- txtProgressBar(style=3)
+      while(!everybodyoutofthepool) {    ###### this loop is really slow - needs a progress bar at least
+        thisregion <- toplace[1]
+        if(length(toplace)>1) {
+          jinregion <- 2
+          for(iinregion in 2:length(toplace)) {
+            thisroute <- detectroute1(start=toplace[1],end=toplace[iinregion],stopiferror=FALSE,algorithm="Dijkstra",rivers=rivers4)
+            num <- num+1  ## 867
+            if(!is.na(thisroute[1])) {
+              thisregion[jinregion] <- toplace[iinregion]
+              jinregion <- jinregion+1
+            }
+            setTxtProgressBar(pb=pb, value=(length(takeout)-length(toplace))/length(takeout)+length(thisregion)/length(takeout))
+          }
+        }
+        regions[[regioni]] <- thisregion
+        regioni <- regioni+1
+        toplace <- toplace[!toplace %in% thisregion]
+        if(length(toplace)==0) everybodyoutofthepool <- T
+        # thisregion <- toplace[1]
+        # if(length(toplace)>1) {
+        #   jinregion <- 2
+        #   toplaceTF <- tocheckTF <- rep(F,length(toplace))
+        #   toplaceTF[1] <- tocheckTF[1] <- T
+        #   while(!all(tocheckTF)) {
+        #     thisroute <- detectroute1(start=toplace[1],end=toplace[which.min(tocheckTF)],stopiferror=FALSE,algorithm="Dijkstra",rivers=rivers4)
+        #     num <- num+1   ## 774
+        #     if(!is.na(thisroute[1])) {
+        #       # thisregion <- c(thisregion,thisroute)
+        #       # toplace <- setdiff(toplace,thisregion)
+        #       toplaceTF[toplace %in% thisroute] <- T
+        #       tocheckTF[toplace %in% thisroute] <- T
+        #     }
+        #     else tocheckTF[which.min(tocheckTF)] <- T
+        #   }
+        # }
+        # thisregion <- toplace[toplaceTF]
+        # regions[[regioni]] <- unique(thisregion)
+        # regioni <- regioni+1
+        # toplace <- toplace[!toplace %in% thisregion]
+        # if(length(toplace)==0) everybodyoutofthepool <- T
+        setTxtProgressBar(pb=pb, value=(length(takeout)-length(toplace))/length(takeout))
       }
-      if(whattodo=="r" | whattodo=="R") {
-        suppressMessages(rivers5 <- trimriver(trim=takeout,rivers=rivers4))
-        dealtwith<-T
-        cat("No unconnected segments detected.",'\n')
-      }
-      if(whattodo=="c" | whattodo=="C") {
-        plot(rivers4,lwd=2)
+        # })  # 11.8   # 13.18 seconds
+      
+      # system.time({
+      # regionvec <- rep(NA,length(takeout))
+      # regioni <- 1
+      # num <- 0
+      # while(any(is.na(regionvec))) {
+      #   isna <- which(is.na(regionvec))
+      #   for(i in isna) {
+      #     if(is.na(regionvec[i])) {
+      #       for(j in isna) {
+      #         if(is.na(regionvec[j])) {
+      #           theroute <- detectroute(start=takeout[i],end=takeout[j],stopiferror=F,algorithm="Dijkstra",rivers=rivers4)
+      #           num <- num+1
+      #           if(!is.na(theroute[1])) {
+      #             regionvec[takeout %in% theroute] <- regioni
+      #           }
+      #         }
+      #       }
+      #     }
+      #     regioni <- regioni+1
+      #   }
+      # }
+      # num
+      # regions <- list()
+      # aai <- 1
+      # for(aa in sort(unique(regionvec))) {
+      #   regions[[aai]] <- takeout[which(regionvec==aa)]
+      #   aai<-aai+1
+      # }
+      # })  ## 12.7 sec
+      
+      cat('\n',length(regions),"regions identified.",'\n')
+      
+      taketheseout <- NULL
+      connectthese <- tothese <- closestptTF <- NULL
+      for(iregions in 1:length(regions)) {
+        cat("Region",iregions,"of",length(regions),"...")
+        zoomtoseg(regions[[iregions]],rivers=rivers4)
+        highlightseg(regions[[iregions]],rivers=rivers4,add=T)
         topologydots(rivers4,add=T)
-        connect1 <- as.numeric(readline("Enter the number of the segment you'd like to connect: "))
-        connect2 <- as.numeric(readline("Enter the number of the segment you'd like to connect it to: "))
-        howtodo <- 0
-        while(!any(howtodo==c("c","C","e","E"))) {
-          howtodo <- readline(prompt="Connect at (e)ndpoint or (c)losest point? (e/c) ")
+        
+        whattodo<-0
+        while(!any(whattodo==c("r","R","c","C"))) {
+          whattodo <- readline(prompt="Please select - (r)emove region or (c)onnect segments & check topology (r/c) ")
         }
-        if(any(howtodo==c("c","C"))) closestpt <- T
-        if(any(howtodo==c("e","E"))) closestpt <- F
-        if(closestpt) cat('\n',"Connecting and calculating new segment splits...",'\n')
-        if(!closestpt) cat('\n',"Connecting...",'\n')
-        suppressMessages(rivers4a <- connectsegs(connect=connect1,connectto=connect2,nearestvert=closestpt,rivers=rivers4))
-        tozoomto <- rbind(rivers4$lines[[connect1]],rivers4$lines[[connect2]])
-        xmin <- min(tozoomto[,1],na.rm=T)
-        xmax <- max(tozoomto[,1],na.rm=T)
-        ymin <- min(tozoomto[,2],na.rm=T)
-        ymax <- max(tozoomto[,2],na.rm=T)
-        plot(rivers4a,xlim=c(xmin,xmax),ylim=c(ymin,ymax),lwd=2)
-        topologydots(rivers=rivers4a,add=TRUE)
-        whatnow <- 0
-        while(!any(whatnow==c("y","Y","n","N"))) {
-          whatnow <- readline(prompt="Accept changes? (y/n) ")
+        if(whattodo=="r" | whattodo=="R") {
+          taketheseout <- c(taketheseout,regions[[iregions]])
         }
-        if(any(whatnow==c("y","Y"))) rivers4 <- rivers4a 
+        if(whattodo=="c" | whattodo=="C") {
+          whatnow <- 0
+          while(!any(whatnow==c("y","Y"))) {
+            connect1 <- as.numeric(readline("Enter the number of the segment you'd like to connect: "))
+            connect2 <- as.numeric(readline("Enter the number of the segment you'd like to connect it to: "))
+            howtodo <- 0
+            while(!any(howtodo==c("c","C","e","E"))) {
+              howtodo <- readline(prompt="Connect at (e)ndpoint or (c)losest point? (e/c) ")
+            }
+            if(any(howtodo==c("c","C"))) closestpt <- T
+            if(any(howtodo==c("e","E"))) closestpt <- F
+            if(closestpt) cat('\n',"Connecting and calculating new segment splits...",'\n')
+            if(!closestpt) cat('\n',"Connecting...",'\n')
+            suppressMessages(rivers4a <- connectsegs_multiple(connect=connect1,connectto=connect2,nearestvert=closestpt,rivers=rivers4))   ##### 
+            tozoomto <- rbind(rivers4$lines[[connect1]],rivers4$lines[[connect2]])
+            xmin <- min(tozoomto[,1],na.rm=T)
+            xmax <- max(tozoomto[,1],na.rm=T)
+            ymin <- min(tozoomto[,2],na.rm=T)
+            ymax <- max(tozoomto[,2],na.rm=T)
+            plot(rivers4a,xlim=c(xmin,xmax),ylim=c(ymin,ymax),lwd=2)
+            topologydots(rivers=rivers4a,add=TRUE)                     # doesn't show newly created segment numbers??  ..........doesn't seem to make new segments ?????
+            whatnow <- 0
+            while(!any(whatnow==c("y","Y","n","N"))) {
+              whatnow <- readline(prompt="Accept changes? (y/n) ")   ########## option to redo when changes are not accepted!
+            }
+            if(any(whatnow==c("n","N"))) {
+              zoomtoseg(regions[[iregions]],rivers=rivers4)
+              highlightseg(regions[[iregions]],rivers=rivers4,add=T)
+              topologydots(rivers4,add=T)
+            }
+          }
+          # if(any(whatnow==c("y","Y"))) {
+            connectthese <- c(connectthese, connect1)
+            tothese <- c(tothese, connect2)
+            closestptTF <- c(closestptTF, closestpt)
+          # }
+        }
       }
+      
+      if(!is.null(connectthese)) suppressMessages(rivers4 <- connectsegs_multiple(connect=connectthese, connectto=tothese, nearestvert=closestptTF, rivers=rivers4))
+      if(!is.null(taketheseout)) suppressMessages(rivers4 <- trimriver(trim=taketheseout, rivers=rivers4))
+      
+      ##################
+      
+      # whattodo<-0
+      # while(!any(whattodo==c("r","R","c","C"))) {
+      #   whattodo <- readline(prompt="Please select - (r)emove unconnected segments or (c)onnect segments & check again (r/c) ")
+      # }
+      # if(whattodo=="r" | whattodo=="R") {
+      #   suppressMessages(rivers5 <- trimriver(trim=takeout,rivers=rivers4))
+      #   dealtwith<-T
+      #   cat("No unconnected segments detected.",'\n')
+      # }
+      # if(whattodo=="c" | whattodo=="C") {
+      #   plot(rivers4,lwd=2)
+      #   topologydots(rivers4,add=T)
+      #   connect1 <- as.numeric(readline("Enter the number of the segment you'd like to connect: "))
+      #   connect2 <- as.numeric(readline("Enter the number of the segment you'd like to connect it to: "))
+      #   howtodo <- 0
+      #   while(!any(howtodo==c("c","C","e","E"))) {
+      #     howtodo <- readline(prompt="Connect at (e)ndpoint or (c)losest point? (e/c) ")
+      #   }
+      #   if(any(howtodo==c("c","C"))) closestpt <- T
+      #   if(any(howtodo==c("e","E"))) closestpt <- F
+      #   if(closestpt) cat('\n',"Connecting and calculating new segment splits...",'\n')
+      #   if(!closestpt) cat('\n',"Connecting...",'\n')
+      #   suppressMessages(rivers4a <- connectsegs(connect=connect1,connectto=connect2,nearestvert=closestpt,rivers=rivers4))
+      #   tozoomto <- rbind(rivers4$lines[[connect1]],rivers4$lines[[connect2]])
+      #   xmin <- min(tozoomto[,1],na.rm=T)
+      #   xmax <- max(tozoomto[,1],na.rm=T)
+      #   ymin <- min(tozoomto[,2],na.rm=T)
+      #   ymax <- max(tozoomto[,2],na.rm=T)
+      #   plot(rivers4a,xlim=c(xmin,xmax),ylim=c(ymin,ymax),lwd=2)
+      #   topologydots(rivers=rivers4a,add=TRUE)
+      #   whatnow <- 0
+      #   while(!any(whatnow==c("y","Y","n","N"))) {
+      #     whatnow <- readline(prompt="Accept changes? (y/n) ")
+      #   }
+      #   if(any(whatnow==c("y","Y"))) rivers4 <- rivers4a 
+      # }
     }
-    if(is.null(takeout)) {
+    if(length(takeout)==0) {
       rivers5<-rivers4
       dealtwith<-T
-      cat("No unconnected segments detected.",'\n')
+      cat('\n',"No unconnected segments detected.",'\n')
     }
   }
   
@@ -334,7 +531,7 @@ cleanup <- function(rivers) {
   }
   
   cat('\n',"Checking for braiding...",'\n')
-  routes<-checkbraidedTF(rivers5,toreturn="routes")
+  routes<-checkbraidedTF1(rivers5,toreturn="routes")           ##################################################################
   rivers5$braided <- !is.null(routes)
   braided <- rivers5$braided
   rivers6<-rivers5
@@ -368,7 +565,7 @@ cleanup <- function(rivers) {
       checkagain<-0
       while(!any(checkagain==c("y","Y","n","N"))) checkagain<-readline(prompt="Re-check for braiding? (y/n) ")
       if(checkagain=="Y" | checkagain=="y") {
-        routes<-checkbraidedTF(rivers6,toreturn="routes")
+        routes<-checkbraidedTF1(rivers6,toreturn="routes")           ##################################################################
         rivers6$braided <- !is.null(routes)
         braided <- rivers6$braided
         if(braided) cat('\n',"Braiding still detected.",'\n')
@@ -492,6 +689,90 @@ connectsegs <- function(connect,connectto,nearestvert=FALSE,rivers,calcconnectio
 
 
 
+connectsegs_multiple <- function(connect,connectto,nearestvert,rivers,calcconnections=TRUE) {
+  pdist2 <- function(p1,p2mat) {
+    dist <- sqrt((p1[1]-p2mat[,1])^2 + (p1[2]-p2mat[,2])^2)
+    return(dist)
+  }
+  for(ientry in 1:length(connect)) {
+    if(length(whoconnected(connect[ientry],rivers))>0){
+      if(any(whoconnected(connect[ientry],rivers)==connectto[ientry])) cat('\n',"Segments",connect[ientry],"and",connectto[ientry],"are already connected.")
+    }
+    l1 <- dim(rivers$lines[[connect[ientry]]])[1]
+    l2 <- dim(rivers$lines[[connectto[ientry]]])[1]
+    if(!nearestvert[ientry]) {
+      dists <- pdist(rivers$lines[[connect[ientry]]][1,],rivers$lines[[connectto[ientry]]][1,])
+      dists[2] <- pdist(rivers$lines[[connect[ientry]]][1,],rivers$lines[[connectto[ientry]]][l2,])
+      dists[3] <- pdist(rivers$lines[[connect[ientry]]][l1,],rivers$lines[[connectto[ientry]]][1,])
+      dists[4] <- pdist(rivers$lines[[connect[ientry]]][l1,],rivers$lines[[connectto[ientry]]][l2,])
+      if(min(dists)==dists[1]) {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connectto[ientry]]][1,],rivers$lines[[connect[ientry]]])
+      }
+      if(min(dists)==dists[2]) {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connectto[ientry]]][l2,],rivers$lines[[connect[ientry]]])
+      }
+      if(min(dists)==dists[3]) {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connect[ientry]]],rivers$lines[[connectto[ientry]]][1,])
+      }
+      if(min(dists)==dists[4]) {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connect[ientry]]],rivers$lines[[connectto[ientry]]][l2,])
+      }
+      
+      rivers$sp@lines[[rivers$lineID[connect[ientry],2]]]@Lines[[rivers$lineID[connect[ientry],3]]]@coords <- rivers$lines[[connect[ientry]]]
+      rivers$lengths[[connect[ientry]]] <- rivers$lengths[[connect[ientry]]]+min(dists)
+      
+    }
+    
+    if(nearestvert[ientry]) {
+      # dbeg <- dend <- 100000*max(rivers$lengths)
+      # whichbeg <- whichend <- NA
+      # for(i in 1:l2) {
+      #   if(pdist(rivers$lines[[connect[ientry]]][1,],rivers$lines[[connectto[ientry]]][i,])<=dbeg) {    ##### i feel like this should be vectorized
+      #     dbeg <- pdist(rivers$lines[[connect[ientry]]][1,],rivers$lines[[connectto[ientry]]][i,])
+      #     whichbeg <- i
+      #   }
+      #   if(pdist(rivers$lines[[connect[ientry]]][l1,],rivers$lines[[connectto[ientry]]][i,])<=dend) {
+      #     dend <- pdist(rivers$lines[[connect[ientry]]][l1,],rivers$lines[[connectto[ientry]]][i,])
+      #     whichend <- i
+      #   }
+      # }
+      dbeg <- pdist2(rivers$lines[[connect[ientry]]][1,],rivers$lines[[connectto[ientry]]])
+      dend <- pdist2(rivers$lines[[connect[ientry]]][l1,],rivers$lines[[connectto[ientry]]])
+      whichbeg <- which.min(dbeg)
+      whichend <- which.min(dend)
+      if(min(dbeg,na.rm=T) < min(dend,na.rm=T)) {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connectto[ientry]]][whichbeg,],rivers$lines[[connect[ientry]]])
+      }
+      else {
+        rivers$lines[[connect[ientry]]] <- rbind(rivers$lines[[connect[ientry]]],rivers$lines[[connectto[ientry]]][whichend,])
+      }
+      
+      rivers$sp@lines[[rivers$lineID[connect[ientry],2]]]@Lines[[rivers$lineID[connect[ientry],3]]]@coords <- rivers$lines[[connect[ientry]]]
+      rivers$lengths[[connect[ientry]]] <- rivers$lengths[[connect[ientry]]]+min(c(dbeg,dend))
+      
+    }
+  }
+  
+  # updating the connectivity matrix
+  if(any(nearestvert)) suppressMessages(rivers <- splitsegments_append(rivers,splitthese=connectto,splitthemat=connect,one2one=T))   ##### make this the multiple one
+  else {
+    # length <- length(rivers$lines)
+    if(calcconnections) rivers$connections <- calculateconnections(lines=rivers$lines,tolerance=rivers$tolerance)
+    if(any(rivers$connections %in% 5:6)) rivers$braided <- TRUE
+  }
+  
+  if(!is.null(rivers$segroutes)) {
+    rivers <- buildsegroutes(rivers,lookup=F)
+  }
+  rivers <- addcumuldist(rivers)
+  if(!is.null(rivers$distlookup)) rivers <- buildlookup(rivers)
+  
+  message("Note: any point data already using the input river network must be re-transformed to river coordinates using xy2segvert() or ptshp2segvert().")
+  return(rivers)
+}
+
+
+
 #' Remove Segments that are Smaller than the Connectivity Tolerance
 #' @description Automatically detects and removes segments with total displacement (straight-line distance between endpoints) less than the connectivity tolerance.  These segments do not serve any real purpose, are bypassed in routes, and cannot be dissolved.
 #' @param rivers The river network object to use.
@@ -515,7 +796,7 @@ removemicrosegs <- function(rivers) {
     for(jj in connectedto) {
       for(jjj in connectedto) {
         if(jj!=jjj & !any(whoconnected(seg=jj,rivers=rivers)==jjj)) {
-          rivers <- connectsegs(connect=jj,connectto=jjj,rivers=rivers,calcconnections=F)
+          rivers <- suppressMessages(connectsegs(connect=jj,connectto=jjj,rivers=rivers,calcconnections=F))
         }
       }
     }